Move parsing model
