+++ cg/apps/sequencing_metrics_parser/models/__init__.py
@@ -1,0 +1,1 @@
+from .bcl2fastq_metrics import SequencingMetricsForLaneAndSample
